Added test for missing VHD and assert VM is absent
Added test for: Error thrown when VM resource's Ensure property equals false #67
It verifies that the Test-TargetResource returns true if the state is "Absent", the VM is missing, and the VHD doesn't exist. If the VM is missing and we wanted it to be so, we don't care if the VHD exists.